Restructurization step 4
-> Moved into App project LayoutingAlgorithms module
