+++ TagCloudContainerTask/TagCloudContainerTask.sln
@@ -5,8 +5,6 @@ VisualStudioVersion = 16.0.31729.503
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ConsoleClient", "ConsoleClient\ConsoleClient.csproj", "{714B2344-1B3F-4BE1-B878-E783E6CDDD07}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TagCloud", "TagCloud\TagCloud.csproj", "{31ACEB8C-2408-42BD-A5CC-97D0F8D9DC4D}"
-EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "GuiClient", "GuiClient\GuiClient.csproj", "{D0D32194-E55C-4C5D-A671-B9A514F71AAB}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "App", "App\App.csproj", "{7561D921-776D-49A5-9221-E52030A61402}"
@@ -21,10 +19,6 @@ Global
 		{714B2344-1B3F-4BE1-B878-E783E6CDDD07}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{714B2344-1B3F-4BE1-B878-E783E6CDDD07}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{714B2344-1B3F-4BE1-B878-E783E6CDDD07}.Release|Any CPU.Build.0 = Release|Any CPU
-		{31ACEB8C-2408-42BD-A5CC-97D0F8D9DC4D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{31ACEB8C-2408-42BD-A5CC-97D0F8D9DC4D}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{31ACEB8C-2408-42BD-A5CC-97D0F8D9DC4D}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{31ACEB8C-2408-42BD-A5CC-97D0F8D9DC4D}.Release|Any CPU.Build.0 = Release|Any CPU
 		{D0D32194-E55C-4C5D-A671-B9A514F71AAB}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{D0D32194-E55C-4C5D-A671-B9A514F71AAB}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{D0D32194-E55C-4C5D-A671-B9A514F71AAB}.Release|Any CPU.ActiveCfg = Release|Any CPU
